sharepoint.py fix
